zephyr-doc-theme: restore breadcrumb nav header
breadcrumb navigation header was in the previous theme but was removed in the
new website. This provides a valuable navigation context reference. include
the breadcrumb navigation in the layout.html, and tweak the CSS to make room
between the site header and main content div (tweaks to custom.css and
main.css)

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>